+++ CHANGELOG.md
@@ -7,6 +7,7 @@ We follow [Semantic Versions](https://semver.org/).
 - Relax version constraint for python to allow 4.0 and newer
 - Make settings pulling for django remote db dumps case insensitive
 - Add support for database urls in django remote db dumps invocations
+- Add invocation for opening shell in k8s db pod
 
 ## 1.6.0